Create SEGeneralRealmObject.kt to realm

+++ app/src/main/java/jp/co/my/mysen/view_model/SEBaseRealmViewModel.kt
@@ -1,11 +1,13 @@
 package jp.co.my.mysen.view_model
 
+import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.liveData
 import io.realm.Realm
 import io.realm.kotlin.where
 import jp.co.my.mysen.realm.SECountryRealmObject
+import jp.co.my.mysen.realm.SEGeneralRealmObject
 import kotlinx.coroutines.Dispatchers
 import okhttp3.OkHttpClient
 import okhttp3.Request
@@ -19,18 +21,27 @@ class SEBaseRealmViewModel: ViewModel() {
 
     fun createBaseRealms(): LiveData<Boolean> {
         return liveData(Dispatchers.IO) {
-            val response = httpGet("https://script.google.com/macros/s/AKfycbzVk_CvxzpIAxXhx5lroyDzayI6sV4TLoLiHYU-CgwbuHKbPTM/exec?sheet=country")
-            response?.also {
-                val realm = Realm.getDefaultInstance()
-                realm.where<SECountryRealmObject>().findAll().deleteAllFromRealm()
-
-                realm.executeTransaction {
-                    it.createAllFromJson(SECountryRealmObject::class.java, response)
-                }
-                emit(true)
-            } ?: run {
+            val countryJson = httpGet("https://script.google.com/macros/s/AKfycbzVk_CvxzpIAxXhx5lroyDzayI6sV4TLoLiHYU-CgwbuHKbPTM/exec?sheet=country")
+            val generalJson = httpGet("https://script.google.com/macros/s/AKfycbzVk_CvxzpIAxXhx5lroyDzayI6sV4TLoLiHYU-CgwbuHKbPTM/exec?sheet=general")
+            if (countryJson == null || generalJson == null) {
                 emit(false)
+                return@liveData
             }
+            val realm = Realm.getDefaultInstance()
+            realm.beginTransaction()
+            // 書き込み
+            realm.where<SECountryRealmObject>().findAll().deleteAllFromRealm()
+            realm.where<SEGeneralRealmObject>().findAll().deleteAllFromRealm()
+            realm.createAllFromJson(SECountryRealmObject::class.java, countryJson)
+            realm.createAllFromJson(SEGeneralRealmObject::class.java, generalJson)
+            // リレーションシップ登録
+            val countries = realm.where<SECountryRealmObject>().sort("id").findAll()
+            val generals = realm.where<SEGeneralRealmObject>().findAll()
+            generals.forEach {
+                it.country = countries[it.country_id.toInt()]
+            }
+            realm.commitTransaction()
+            emit(true)
         }
     }
